End the practice of retaining the SchemaStore in the Schema.
It invites breaking the strict 'arrow flow' of (1) build an immutable SchemaStore (2) build immutable Schema from SchemasStore (3) validate from Schema.

This architecture is acting to keep some order in what is an increasingly complex system. I would like to prevent the Validator from having access (direct or indirect) to the SchemaStore for as long as I can. Having the Schema hold on to the SchemaStore invites future maintainers to do things with it in Schema's getters or even vend with it.