Stop using schemaObject for non-debugging purposes.
I wrote that schemaObject was for debugging only then broke this rule in isFalse.

It's a good idea not to do anything with schemaObject after construction because it's a slippery slop to a design that does some things on construction and others on demand (e.g. in getters); bypassing the layers in the design for no good reason and increasing the confusion of people trying to work with the code.

I do vend it in the API, because I imagine users might find that useful, and I'm only giving them back their own data.

I have thought about a design that is the reverse; only schemaObject is kept and all calculation is done on demand in getters.