Changes to how and where example validation is done.
Now example validation must be requested explicitly after Schema creation. It felt overly restrictive to refuse to build a Schema if the examples didn't validate; plus it required making and using a default validator in the Schema constructor which spoiled the demarcation and might produce unexpected results as the default validator is not as standard-compliant as possible.

Example validation is tricky to do as expected as it must validate examples in all subschemas. To facilitate this the Schema now registers and vends its subschemas, and I tightened the definition of subschema, breaking out a new concept of child schema.

Updated the test data accordingly as known example failures are now registered.